+++ src/main/java/com/gramo/gramo/service/picu/PicuServiceImpl.java
@@ -27,9 +27,9 @@ public class PicuServiceImpl implements PicuService{
 
     @Override
     public List<PicuContentResponse> getPicu(LocalDate date) {
-//        if(!authenticationFacade.isLogin()) {
-//            throw new LoginException();
-//        }
+        if(!authenticationFacade.isLogin()) {
+            throw new LoginException();
+        }
 
         List<Picu> picuList = picuRepository.findAllByDateOrderByDateDesc(date);
         List<PicuContentResponse> picuContentResponses = new ArrayList<>();
@@ -50,13 +50,13 @@ public class PicuServiceImpl implements PicuService{
 
     @Override
     public void createPicu(PicuRequest request) {
-//        if(!authenticationFacade.isLogin()) {
-//            throw new LoginException();
-//        }
+        if(!authenticationFacade.isLogin()) {
+            throw new LoginException();
+        }
 
         picuRepository.save(
                 Picu.builder()
-                        .userEmail("student@dsm.hs.kr")
+                        .userEmail(authenticationFacade.getUserEmail())
                         .description(request.getDescription())
                         .date(request.getDate())
                         .build()
@@ -65,16 +65,19 @@ public class PicuServiceImpl implements PicuService{
 
     @Override
     public void deletePicu(Long picuId) {
-//        if(!authenticationFacade.isLogin()) {
-//            throw new LoginException();
-//        }
+        if(!authenticationFacade.isLogin()) {
+            throw new LoginException();
+        }
 
         Picu picu = picuRepository.findById(picuId)
                 .orElseThrow(PicuNotFoundException::new);
 
-//        if(!picu.getUserEmail().equals(authenticationFacade.getUserEmail())) {
-//            throw new PermissionMismatchException();
-//        }
+        System.out.println(picu.getUserEmail());
+        System.out.println(authenticationFacade.getUserEmail());
+
+        if(!picu.getUserEmail().equals(authenticationFacade.getUserEmail())) {
+            throw new PermissionMismatchException();
+        }
 
         picuRepository.delete(picu);
     }
